Unit test for presenter
